move draggable view to draggableview.swift
