+++ ImageConverter/Image Conversion/NSImage+Conversion.swift
